server/requirements.txt
django
django-cors-headers
google-generativeai
whitenoise
python-dotenv
djangorestframework